requirements.txt
openai==1.12.0
python-dotenv==1.0.0
boto3==1.34.34
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.6
langchain==0.1.9
langchain-openai==0.0.5 